feat: first example of useSWR on AgendaPage
