+++ agir/events/components/agendaPage/AgendaPage.js
@@ -5,11 +5,13 @@ import style from "@agir/front/genericComponents/_variables.scss";
 import Layout from "@agir/front/dashboardComponents/Layout";
 import Agenda from "@agir/events/agendaPage/Agenda";
 
-const AgendaPage = (props) => (
-  <Layout active="events" smallBackgroundColor={style.black25} hasBanner>
-    <Agenda {...props} />
-  </Layout>
-);
+const AgendaPage = (props) => {
+  return (
+    <Layout active="events" smallBackgroundColor={style.black25} hasBanner>
+      <Agenda {...props} />
+    </Layout>
+  );
+};
 
 export default AgendaPage;
