+++ agir/front/views.py
@@ -132,112 +132,9 @@ class FullSupportGroupView(SoftLoginRequiredMixin, ReactSingleObjectView):
             "groupSuggestions": person_groups,
         }
 
-
-class AgendaView(SoftLoginRequiredMixin, ReactSerializerBaseView):
+class AgendaView(SoftLoginRequiredMixin, ReactBaseView):
     bundle_name = "events/agendaPage"
     serializer_class = EventSerializer
-
-    def get(self, request, *args, **kwargs):
-        person = request.user.person
-
-        if person.coordinates_type is None:
-            geocode_person.delay(person.pk)
-
-        return super().get(request, *args, **kwargs)
-
-    def get_export_data(self):
-        person = self.request.user.person
-        queryset = Event.objects.with_serializer_prefetch(person)
-        if person.is_2022_only:
-            queryset = queryset.is_2022()
-
-        rsvped_events = (
-            queryset.upcoming()
-            .filter(Q(attendees=person) | Q(organizers=person))
-            .order_by("start_time", "end_time")
-        ).distinct()
-
-        groups_events = (
-            queryset.upcoming()
-            .filter(organizers_groups__in=person.supportgroups.all())
-            .distinct()
-        )
-
-        organized_events = (
-            queryset.past()
-            .filter(organizers=person)
-            .distinct()
-            .order_by("-start_time")[:10]
-        )
-
-        past_events = (
-            queryset.past()
-            .filter(
-                Q(rsvps__person=person)
-                | Q(organizers_groups__in=person.supportgroups.all())
-            )
-            .distinct()
-            .order_by("-start_time")[:10]
-        )
-
-        other_events = groups_events.union(organized_events, past_events)
-
-        if person.coordinates is not None:
-            near_events = (
-                queryset.upcoming()
-                .filter(
-                    start_time__lt=timezone.now() + timedelta(days=30),
-                    do_not_list=False,
-                )
-                .exclude(pk__in=groups_events)
-                .annotate(distance=Distance("coordinates", person.coordinates))
-                .order_by("distance")[:10]
-            )
-
-            other_events = (
-                queryset.filter(
-                    pk__in=[e.pk for e in near_events] + [e.pk for e in other_events]
-                )
-                .annotate(distance=Distance("coordinates", person.coordinates))
-                .order_by("start_time")
-                .distinct()
-            )
-        else:
-            other_events = other_events.order_by("start_time")
-
-        fields = [
-            "id",
-            "name",
-            "participantCount",
-            "illustration",
-            "hasSubscriptionForm",
-            "startTime",
-            "endTime",
-            "location",
-            "isOrganizer",
-            "rsvp",
-            "canRSVP",
-            "is2022",
-            "routes",
-            "groups",
-            "distance",
-            "compteRendu",
-        ]
-
-        return {
-            "rsvped": self.serializer_class(
-                instance=rsvped_events,
-                many=True,
-                context={"request": self.request},
-                fields=fields,
-            ).data,
-            "others": self.serializer_class(
-                instance=other_events,
-                many=True,
-                context={"request": self.request},
-                fields=fields,
-            ).data,
-        }
 
 
 class MyGroupsView(SoftLoginRequiredMixin, ReactListView):
