+++ package.json
@@ -68,6 +68,7 @@
     "short-uuid": "^4.1.0",
     "styled-components": "^5.2.0",
     "swiper": "^6.3.5",
+    "swr": "^0.3.10",
     "tinymce": "^4.9.11",
     "tinymce-i18n": "^17.12.31",
     "turbolinks": "^5.2.0"
